Split amp-inline-gallery component files
