+++ extensions/amp-inline-gallery/1.0/amp-inline-gallery.js
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import * as Preact from '../../../src/preact';
 import {
   AmpInlineGalleryPagination,
   TAG as PAGINATION_TAG,
@@ -23,28 +22,16 @@ import {
   AmpInlineGalleryThumbnails,
   TAG as THUMBNAILS_TAG,
 } from './amp-inline-gallery-thumbnails';
-import {CarouselContextProp} from '../../amp-base-carousel/1.0/carousel-props';
-import {InlineGallery} from './inline-gallery';
+import {BaseElement} from './base-element';
 import {Layout} from '../../../src/layout';
 import {CSS as PAGINATION_CSS} from '../../../build/amp-inline-gallery-pagination-1.0.css';
-import {PreactBaseElement} from '../../../src/preact/base-element';
-import {dict} from '../../../src/utils/object';
 import {isExperimentOn} from '../../../src/experiments';
-import {setProp} from '../../../src/context';
-import {useContext, useLayoutEffect} from '../../../src/preact';
 import {pureUserAssert as userAssert} from '../../../src/core/assert';
 
 /** @const {string} */
 const TAG = 'amp-inline-gallery';
 
-class AmpInlineGallery extends PreactBaseElement {
-  /** @override */
-  init() {
-    return dict({
-      'children': <ContextExporter shimDomElement={this.element} />,
-    });
-  }
-
+class AmpInlineGallery extends BaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -56,17 +43,6 @@ class AmpInlineGallery extends PreactBaseElement {
   }
 }
 
-/** @override */
-AmpInlineGallery['Component'] = InlineGallery;
-
-/** @override */
-AmpInlineGallery['detached'] = true;
-
-/** @override */
-AmpInlineGallery['props'] = {
-  'loop': {attr: 'loop', type: 'boolean'},
-};
-
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpInlineGallery);
   AMP.registerElement(
@@ -76,17 +52,3 @@ AMP.extension(TAG, '1.0', (AMP) => {
   );
   AMP.registerElement(THUMBNAILS_TAG, AmpInlineGalleryThumbnails);
 });
-
-/**
- * @param {!SelectorDef.OptionProps} props
- * @return {PreactDef.Renderable}
- */
-function ContextExporter({shimDomElement}) {
-  // Consume the `CarouselContext` produced by the `InlineGallery` component
-  // and propagate it as a context prop.
-  const context = useContext(CarouselContextProp.type);
-  useLayoutEffect(() => {
-    setProp(shimDomElement, CarouselContextProp, ContextExporter, context);
-  }, [shimDomElement, context]);
-  return <></>;
-}
