+++ extensions/amp-inline-gallery/1.0/base-element.js
@@ -1,0 +1,57 @@
+/**
+ * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as Preact from '../../../src/preact';
+import {CarouselContextProp} from '../../amp-base-carousel/1.0/carousel-props';
+import {InlineGallery} from './component';
+import {PreactBaseElement} from '../../../src/preact/base-element';
+import {dict} from '../../../src/utils/object';
+import {setProp} from '../../../src/context';
+import {useContext, useLayoutEffect} from '../../../src/preact';
+
+export class BaseElement extends PreactBaseElement {
+  /** @override */
+  init() {
+    return dict({
+      'children': <ContextExporter shimDomElement={this.element} />,
+    });
+  }
+}
+
+/** @override */
+BaseElement['Component'] = InlineGallery;
+
+/** @override */
+BaseElement['detached'] = true;
+
+/** @override */
+BaseElement['props'] = {
+  'loop': {attr: 'loop', type: 'boolean'},
+};
+
+/**
+ * @param {!SelectorDef.OptionProps} props
+ * @return {PreactDef.Renderable}
+ */
+function ContextExporter({shimDomElement}) {
+  // Consume the `CarouselContext` produced by the `InlineGallery` component
+  // and propagate it as a context prop.
+  const context = useContext(CarouselContextProp.type);
+  useLayoutEffect(() => {
+    setProp(shimDomElement, CarouselContextProp, ContextExporter, context);
+  }, [shimDomElement, context]);
+  return <></>;
+}
