+++ extensions/amp-inline-gallery/1.0/component.js
