+++ extensions/amp-inline-gallery/1.0/test/test-inline-gallery.js
@@ -16,7 +16,7 @@
 
 import * as Preact from '../../../../src/preact';
 import {BaseCarousel} from '../../../amp-base-carousel/1.0/base-carousel';
-import {InlineGallery} from '../inline-gallery';
+import {InlineGallery} from '../component';
 import {Pagination} from '../pagination';
 import {Thumbnails} from '../thumbnails';
 import {mount} from 'enzyme';
